Add files via upload
Final keyword is used:
it can be used with variable, method and class.
i) if we make variable final, the variable will be constant won't be able to change the value of that variable. final int num=10;
ii) if you make the method final. you are not allow to override the method again. eg. public final void show(){}
iii) Once the class gets final we can't inherit that class. eg. final class Parent{}